+++ changelog.md
@@ -14,11 +14,29 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ### Changed
 
-- Added better detection of user data directory
+-
 
 ### Fixed
 
 -
+
+---
+
+## [v0.2.0] - 2022/07/13
+
+### Added
+
+- Setting to control minimum time delay between auto-trim (#1)
+- Option to skip code blocks (#4)
+
+### Changed
+
+- Mechanism used for auto-trimming
+- 'Multiple Characters' to only trim inline characters (#2)
+
+### Fixed
+
+- 'Multiple Characters' trimming within tables (#3)
 
 ---
 
@@ -29,4 +47,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - Initial commit!
 
 [Unreleased]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/master...develop
-[v0.1.0]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/9342c0e...v0.1.0
+[v0.2.0]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/v0.1.0...v0.2.0
+[v0.1.0]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/a53bdb3...v0.1.0